requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
strawberry-graphql==0.253.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30